refactor: 패키지 구조 개선 및 불필요 주석 제거
