--- src/main/java/blackjack/domain/card/CardsGenerator.java
@@ -1,8 +1,0 @@
-package blackjack.domain.card;
-
-import java.util.List;
-
-public interface CardsGenerator {
-
-    List<Card> generate();
-}
