+++ src/main/java/blackjack/domain/card/generator/CardsGenerator.java
@@ -1,0 +1,10 @@
+package blackjack.domain.card.generator;
+
+import blackjack.domain.card.Card;
+import java.util.List;
+
+@FunctionalInterface
+public interface CardsGenerator {
+
+    List<Card> generate();
+}
