+++ src/test/java/blackjack/domain/card/DenominationTest.java
@@ -1,9 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import static blackjack.domain.card.Denomination.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.card.Denomination;
+import blackjack.domain.DenominationArgumentsProvider;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
