+++ src/test/java/blackjack/domain/participant/NameTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
@@ -10,6 +10,12 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 public class NameTest {
 
+    @DisplayName("이름이 정상적으로 생성된다.")
+    @Test
+    void 이름_생성_정상() {
+        assertDoesNotThrow(() -> new Name("mat"));
+    }
+
     @DisplayName("이름은 공백이 아니어야한다.")
     @ParameterizedTest
     @ValueSource(strings = {"", " ", "   "})
@@ -17,11 +23,5 @@ public class NameTest {
         assertThatThrownBy(() -> new Name(value))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("이름은 공백이 아니어야합니다.");
-    }
-
-    @DisplayName("이름이 정상적으로 생성된다.")
-    @Test
-    void 이름_생성_정상() {
-        assertDoesNotThrow(() -> new Name("mat"));
     }
 }
